bridge/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
paho-mqtt>=2.0.0
python-multipart>=0.0.6
